requirement.txt
openai
requests
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-experimental
pandas
tabulate
streamlit